requirements.txt
amqp==2.6.1
appdirs==1.4.4
billiard==3.6.3.0
black==20.8b1
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.60.1
gunicorn==20.0.4
h11==0.9.0
httptools==0.1.1
idna==2.10
kombu==4.6.11
mypy-extensions==0.4.3
pathspec==0.8.0
pydantic==1.6.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
starlette==0.13.6
toml==0.10.1
transitions==0.8.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
vine==1.3.0
websockets==8.1
